Consolidate ColorWell thumb elements
