+++ packages/colorpickers/src/styled/ColorPicker/StyledColorWellWrapper.ts
