+++ packages/colorpickers/src/styled/index.ts
@@ -10,11 +10,10 @@ export * from './ColorPicker/StyledHue';
 export * from './ColorPicker/StyledAlpha';
 export * from './ColorPicker/StyledPreview';
 export * from './ColorPicker/StyledColorWell';
-export * from './ColorPicker/StyledSaturationWrapper';
+export * from './ColorPicker/StyledColorWellWrapper';
 export * from './ColorPicker/StyledSaturationWhite';
 export * from './ColorPicker/StyledSaturationBlack';
-export * from './ColorPicker/StyledSaturationPointer';
-export * from './ColorPicker/StyledSaturationCircle';
+export * from './ColorPicker/StyledColorWellThumb';
 export * from './ColorPicker/StyledSliderGroup';
 export * from './ColorPicker/StyledHexField';
 export * from './ColorPicker/StyledInput';
